requirements.txt
Pillow
djangorestframework